requirements.txt
ultralytics>=8.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
opencv-python>=4.8.0
python-telegram-bot>=20.0
websockets>=11.0
python-dotenv>=1.0.0
pillow>=10.0.0
